Refactor: lib/schedule*
