+++ lib/schedule.js
@@ -1,3 +1,20 @@
+/*
+ * This file is part of the Companion project
+ * Copyright (c) 2018 Bitfocus AS
+ * Authors: William Viker <william@bitfocus.io>, Håkon Nessjøen <haakon@bitfocus.io>
+ *
+ * This program is free software.
+ * You should have received a copy of the MIT licence as well as the Bitfocus
+ * Individual Contributor License Agreement for companion along with
+ * this program.
+ *
+ * You can be released from the requirements of the license by purchasing
+ * a commercial license. Buying such a license is mandatory as soon as you
+ * develop commercial activities involving the Companion software without
+ * disclosing the source code of your own applications.
+ *
+ */
+
 const fs = require('fs');
 const debug = require('debug')('lib/instance');
 
@@ -29,106 +46,35 @@ class scheduler {
 	}
 
 	/**
-	 *
-	 * @param {SocketIO} socket
-	 * @param {pluginCallback} cb
+	 * Call an action
+	 * This method is called from within a plugin and sends the id, the scheduler determines what should then happen
+	 * @param {number} id
 	 */
-	send_plugins(socket, cb) {
-		/**
-		 * @callback pluginCallback
-		 * @param {Object[]}
-		 */
-		cb(this.plugins.map(p => p.front_end()));
-	}
+	action(id) {
+		const event = this.config.find(x => x.id === id);
 
-	/**
-	 * Loads plugins from the schedule plugin directory
-	 */
-	load_plugins() {
-		const path = require('app-root-path') + '/lib/schedule';
-		const plugins_folder = fs.readdirSync(path);
-
-		plugins_folder.forEach(p => {
-			if (p === 'plugin_base.js' || p.match(/\.js$/) === null) {
-				return;
-			}
-
-			try {
-				const plugin = require(path + '/' + p);
-				this.plugins.push(new plugin(this));
-			} catch (e) {
-				debug(e);
-				this.system.emit('log', 'scheduler', 'error', `Error loading plugin ${p}`);
-			}
-		});
-	}
-
-	/**
-	 * Initialize the configuration and start any schedules
-	 * @param {?Object[]} res
-	 */
-	init_config(res) {
-		if (res === undefined || res === null) {
-			res = [];
-			this.system.emit('db_set', 'scheduler', res);
-		}
-
-		this.config = res;
-
-		this.start_schedule();
-	}
-
-	/**
-	 * Starts all event schedules
-	 */
-	start_schedule() {
-		this.config.forEach(i => this.event_watch(i));
-	}
-
-	/**
-	 * Gets the plugin from a request type
-	 * @param {string} type
-	 * @return {?Object}
-	 */
-	get_plugin(type) {
-		let plugin = this.plugins.find(p => p.type === type);
-		if (!plugin) {
-			this.system.emit('log', 'scheduler', 'error', 'Plugin not loaded.');
-			return null;
-		}
-		return plugin;
-	}
-
-	/**
-	 * Register or unregister an event from being watched
-	 * @param {Object} config
-	 * @param {boolean} add Add or remove the event from the watch schedule
-	 */
-	event_watch(config, add = true) {
-		let plugin = this.get_plugin(config.type);
-		if (!plugin) {
+		if (!event) {
+			this.system.emit('log', 'scheduler', 'error', 'Could not find configuration for action.')
 			return;
 		}
 
-		if (!add) {
-			this.system.emit('log', 'scheduler', 'info', 'Removing scheduled event.');
-			plugin.remove(config.id);
-		} else if (config.disabled === false) {
-			this.system.emit('log', 'scheduler', 'info', 'Adding scheduled event.');
-			plugin.add(config.id, config);
+		const [bank, button] = this._get_bank_button(event.button);
+
+		this.system.emit('log', 'scheduler', 'info', `Push button ${bank}.${button} via <code>${event.title}</code>`);
+		this.system.emit('bank_pressed', bank, button, true, 'scheduler');
+
+		setTimeout(() => {
+			this.system.emit('bank_pressed', bank, button, false, 'scheduler');
+			this.system.emit('log', 'scheduler', 'info', `Release button ${bank}.${button} via <code>${event.title}</code>`);
+		}, this.btn_release_time);
+
+		// Update the last run
+		event.last_run = new Date();
+		this.save_to_db();
+
+		if (this.io) {
+			this.io.emit('schedule_refresh', this.config);
 		}
-	}
-
-	/**
-	 * Get the bank and button number from a string
-	 * @param {string} button
-	 * @return {number[]}
-	 */
-	_get_bank_button(button) {
-		const bank = parseInt(button);
-		const button_number = parseInt(button.toString().replace(/(.*)\./, ''));
-
-		return [bank, button_number];
 	}
 
 	/**
@@ -159,25 +105,6 @@ class scheduler {
 	}
 
 	/**
-	 * Get the next unique event ID
-	 * @return {number}
-	 */
-	_get_next_id() {
-		if (this.config.length === 0) {
-			return 1;
-		}
-
-		const cur_max_id = Math.max.apply(Math, this.config.map(i => i.id));
-		if (cur_max_id <= 0 || isNaN(cur_max_id)) {
-			debug('current max id is invalid; this may be a bug or a corruption that may require a reset of the scheduler config');
-			this.system.emit('log', 'scheduler', 'warn', 'Configuration appears to be corrupt.');
-			return 1;
-		} else {
-			return cur_max_id + 1;
-		}
-	}
-
-	/**
 	 * Loads plugin based parameters
 	 * These parameters are dynamic and aren't stored to the database, but are
 	 * needed for the front end. For example, the configuration description.
@@ -186,14 +113,97 @@ class scheduler {
 	 */
 	event_load_type(event) {
 		const plugin = this.get_plugin(event.type);
+
 		if (!plugin) {
 			this.system.emit('log', 'scheduler', 'error', `Could not load plugin type ${event.type}.`);
 			event.config_desc = 'Unknown schedule type.';
-		} else {
+		}
+		else {
 			event.config_desc = plugin.config_desc(event.config);
 		}
 
 		return event;
+	}
+
+	/**
+	 * Register or unregister an event from being watched
+	 * @param {Object} config
+	 * @param {boolean} add Add or remove the event from the watch schedule
+	 */
+	event_watch(config, add = true) {
+		let plugin = this.get_plugin(config.type);
+
+		if (!plugin) {
+			return;
+		}
+
+		if (!add) {
+			this.system.emit('log', 'scheduler', 'info', 'Removing scheduled event.');
+			plugin.remove(config.id);
+		}
+		else if (config.disabled === false) {
+			this.system.emit('log', 'scheduler', 'info', 'Adding scheduled event.');
+			plugin.add(config.id, config);
+		}
+	}
+
+	/**
+	 * Get event index from an event ID
+	 * @param {number} id
+	 * @return {number}
+	 */
+	find_event_idx(id) {
+		return this.config.findIndex(x => x.id === id);
+	}
+
+	/**
+	 * Get the bank and button number from a string
+	 * @param {string} button
+	 * @return {number[]}
+	 */
+	_get_bank_button(button) {
+		const bank = parseInt(button);
+		const button_number = parseInt(button.toString().replace(/(.*)\./, ''));
+
+		return [bank, button_number];
+	}
+
+	/**
+	 * Get the next unique event ID
+	 * @return {number}
+	 */
+	_get_next_id() {
+
+		if (this.config.length === 0) {
+			return 1;
+		}
+
+		const cur_max_id = Math.max.apply(Math, this.config.map(i => i.id));
+
+		if (cur_max_id <= 0 || isNaN(cur_max_id)) {
+			debug('current max id is invalid; this may be a bug or a corruption that may require a reset of the scheduler config');
+			this.system.emit('log', 'scheduler', 'warn', 'Configuration appears to be corrupt.');
+			return 1;
+		}
+		else {
+			return cur_max_id + 1;
+		}
+	}
+
+	/**
+	 * Gets the plugin from a request type
+	 * @param {string} type
+	 * @return {?Object}
+	 */
+	get_plugin(type) {
+		let plugin = this.plugins.find(p => p.type === type);
+
+		if (!plugin) {
+			this.system.emit('log', 'scheduler', 'error', 'Plugin not loaded.');
+			return null;
+		}
+
+		return plugin;
 	}
 
 	/**
@@ -210,79 +220,42 @@ class scheduler {
 	}
 
 	/**
-	 * Get event index from an event ID
-	 * @param {number} id
-	 * @return {number}
+	 * Initialize the configuration and start any schedules
+	 * @param {?Object[]} res
 	 */
-	find_event_idx(id) {
-		return this.config.findIndex(x => x.id === id);
+	init_config(res) {
+
+		if (res === undefined || res === null) {
+			res = [];
+			this.system.emit('db_set', 'scheduler', res);
+		}
+
+		this.config = res;
+
+		this.start_schedule();
 	}
 
 	/**
-	 * Call an action
-	 * This method is called from within a plugin and sends the id, the scheduler determines what should then happen
-	 * @param {number} id
+	 * Loads plugins from the schedule plugin directory
 	 */
-	action(id) {
-		const event = this.config.find(x => x.id === id);
-		if (!event) {
-			this.system.emit('log', 'scheduler', 'error', 'Could not find configuration for action.')
-			return;
-		}
+	load_plugins() {
+		const path = require('app-root-path') + '/lib/schedule';
+		const plugins_folder = fs.readdirSync(path);
 
-		const [bank, button] = this._get_bank_button(event.button);
+		plugins_folder.forEach(p => {
+			if (p === 'common.js' || p.match(/\.js$/) === null) {
+				return;
+			}
 
-		this.system.emit('log', 'scheduler', 'info', `Push button ${bank}.${button} via <code>${event.title}</code>`);
-		this.system.emit('bank_pressed', bank, button, true, 'scheduler');
-
-		setTimeout(() => {
-			this.system.emit('bank_pressed', bank, button, false, 'scheduler');
-			this.system.emit('log', 'scheduler', 'info', `Release button ${bank}.${button} via <code>${event.title}</code>`);
-		}, this.btn_release_time);
-
-		// Update the last run
-		event.last_run = new Date();
-		this.save_to_db();
-
-		if (this.io) {
-			this.io.emit('schedule_refresh', this.config);
-		}
-	}
-
-	/**
-	 * Updates a schedule
-	 * Minor updates and deletions
-	 * @param {SocketIO} socket
-	 * @param {number} id Event ID
-	 * @param {Object} new_data If deleted property is set, event is deleted
-	 * @param {scheduleUpdateCb} cb
-	 */
-	update_schedule(socket, id, new_data, cb) {
-		const idx = this.find_event_idx(id);
-		if (idx === -1) {
-			this.system.emit('log', 'scheduler', 'error', 'Scheduled event could not be found.');
-			return;
-		}
-
-		// Stop watching old event
-		this.event_watch(this.config[idx], false);
-
-		/**
-		 * @callback scheduleUpdateCb
-		 * @param {?Object} event Will return null if event was deleted
-		 */
-		if ('deleted' in new_data) {
-			this.config.splice(idx, 1);
-			cb(null);
-		} else {
-			this.config[idx] = {...this.config[idx], ...new_data};
-			cb(this.event_load_type(this.config[idx]));
-			this.event_watch(this.config[idx]);
-		}
-
-		this.save_to_db();
-
-		socket.broadcast.emit('schedule_refresh', this.config);
+			try {
+				const plugin = require(path + '/' + p);
+				this.plugins.push(new plugin(this));
+			}
+			catch (e) {
+				debug(e);
+				this.system.emit('log', 'scheduler', 'error', `Error loading plugin ${p}`);
+			}
+		});
 	}
 
 	/**
@@ -295,9 +268,11 @@ class scheduler {
 		const clean_data = this.clean_config(new_data);
 
 		let idx = this.find_event_idx(clean_data.id);
+
 		if (idx === -1) {
 			this.config.push(clean_data);
-		} else {
+		}
+		else {
 			// Keep the last run and disabled status from the old config
 			clean_data.last_run = this.config[idx].last_run;
 			clean_data.disabled = this.config[idx].disabled;
@@ -323,6 +298,64 @@ class scheduler {
 	save_to_db() {
 		this.system.emit('db_set', 'scheduler', this.config);
 		this.system.emit('db_save');
+	}
+
+	/**
+	 *
+	 * @param {SocketIO} socket
+	 * @param {pluginCallback} cb
+	 */
+	send_plugins(socket, cb) {
+		/**
+		 * @callback pluginCallback
+		 * @param {Object[]}
+		 */
+		cb(this.plugins.map(p => p.front_end()));
+	}
+
+	/**
+	 * Starts all event schedules
+	 */
+	start_schedule() {
+		this.config.forEach(i => this.event_watch(i));
+	}
+
+	/**
+	 * Updates a schedule
+	 * Minor updates and deletions
+	 * @param {SocketIO} socket
+	 * @param {number} id Event ID
+	 * @param {Object} new_data If deleted property is set, event is deleted
+	 * @param {scheduleUpdateCb} cb
+	 */
+	update_schedule(socket, id, new_data, cb) {
+		const idx = this.find_event_idx(id);
+
+		if (idx === -1) {
+			this.system.emit('log', 'scheduler', 'error', 'Scheduled event could not be found.');
+			return;
+		}
+
+		// Stop watching old event
+		this.event_watch(this.config[idx], false);
+
+		/**
+		 * @callback scheduleUpdateCb
+		 * @param {?Object} event Will return null if event was deleted
+		 */
+		if ('deleted' in new_data) {
+			this.config.splice(idx, 1);
+			cb(null);
+		}
+		else {
+			this.config[idx] = {...this.config[idx], ...new_data};
+			cb(this.event_load_type(this.config[idx]));
+			this.event_watch(this.config[idx]);
+		}
+
+		this.save_to_db();
+
+		socket.broadcast.emit('schedule_refresh', this.config);
 	}
 }
