+++ lib/schedule/companion_instance.js
@@ -1,6 +1,24 @@
-const plugin_base = require('./plugin_base');
+/*
+ * This file is part of the Companion project
+ * Copyright (c) 2018 Bitfocus AS
+ * Authors: William Viker <william@bitfocus.io>, Håkon Nessjøen <haakon@bitfocus.io>
+ *
+ * This program is free software.
+ * You should have received a copy of the MIT licence as well as the Bitfocus
+ * Individual Contributor License Agreement for companion along with
+ * this program.
+ *
+ * You can be released from the requirements of the license by purchasing
+ * a commercial license. Buying such a license is mandatory as soon as you
+ * develop commercial activities involving the Companion software without
+ * disclosing the source code of your own applications.
+ *
+ */
 
-class companion_instance extends plugin_base {
+const common = require('./common');
+
+class companion_instance extends common {
+
 	setup() {
 		const system = this.scheduler.system;
 
@@ -44,6 +62,21 @@ class companion_instance extends plugin_base {
 		];
 	}
 
+	config_desc(config) {
+		const run_conf = this.options[0].choices.find(x => x.id === config.run);
+		let run_time = !run_conf ? 'unknown' : run_conf.label;
+
+		return run_time;
+	}
+
+	get name() {
+		return 'Instance';
+	}
+
+	get type() {
+		return 'instance';
+	}
+
 	/**
 	 * Checks if event needs to run
 	 * @param {string} run Name of run type based on the option choices (ie, start or button_press)
@@ -51,21 +84,6 @@ class companion_instance extends plugin_base {
 	_run(run) {
 		this.watch.filter(x => x.config.run === run)
 			.forEach(x => this.scheduler.action(x.id));
-	}
-
-	get type() {
-		return 'instance';
-	}
-
-	get name() {
-		return 'Instance';
-	}
-
-	config_desc(config) {
-		const run_conf = this.options[0].choices.find(x => x.id === config.run);
-		let run_time = !run_conf ? 'unknown' : run_conf.label;
-
-		return run_time;
 	}
 }
