+++ lib/schedule/interval.js
@@ -1,6 +1,24 @@
-const plugin_base = require('./plugin_base');
+/*
+ * This file is part of the Companion project
+ * Copyright (c) 2018 Bitfocus AS
+ * Authors: William Viker <william@bitfocus.io>, Håkon Nessjøen <haakon@bitfocus.io>
+ *
+ * This program is free software.
+ * You should have received a copy of the MIT licence as well as the Bitfocus
+ * Individual Contributor License Agreement for companion along with
+ * this program.
+ *
+ * You can be released from the requirements of the license by purchasing
+ * a commercial license. Buying such a license is mandatory as soon as you
+ * develop commercial activities involving the Companion software without
+ * disclosing the source code of your own applications.
+ *
+ */
 
-class interval extends plugin_base {
+const common = require('./common');
+
+class interval extends common {
+
 	setup() {
 		this.options = [
 			{
@@ -17,16 +35,32 @@ class interval extends plugin_base {
 		data.interval = setInterval(() => {
 			this.scheduler.action(id);
 		}, parseInt(data.config.seconds) * 1000);
+
 		super.add(id, data);
 	}
 
-	remove(id) {
-		const config = this.watch.find(x => x.id == id);
-		if (config && config.interval) {
-			clearInterval(config.interval);
+	config_desc(config) {
+		let time;
+
+		if (config.seconds >= 3600) {
+			time = `${parseInt(config.seconds / 3600)} hours`;
+		}
+		else if (config.seconds >= 60) {
+			time = `${parseInt(config.seconds / 60)} minutes`;
+		}
+		else {
+			time = `${config.seconds} seconds`;
 		}
 
-		super.remove(id);
+		return `Runs every <strong>${time}</strong>.`;
+	}
+
+	get name() {
+		return 'Time interval';
+	}
+
+	get type() {
+		return 'interval';
 	}
 
 	/**
@@ -39,25 +73,14 @@ class interval extends plugin_base {
 			.forEach(x => this.scheduler.action(x.id));
 	}
 
-	get type() {
-		return 'interval';
-	}
+	remove(id) {
+		const config = this.watch.find(x => x.id == id);
 
-	get name() {
-		return 'Time interval';
-	}
-
-	config_desc(config) {
-		let time;
-		if (config.seconds >= 3600) {
-			time = `${parseInt(config.seconds / 3600)} hours`;
-		} else if (config.seconds >= 60) {
-			time = `${parseInt(config.seconds / 60)} minutes`;
-		} else {
-			time = `${config.seconds} seconds`;
+		if (config && config.interval) {
+			clearInterval(config.interval);
 		}
 
-		return `Runs every <strong>${time}</strong>.`;
+		super.remove(id);
 	}
 }
