+++ lib/schedule/time_of_day.js
@@ -1,7 +1,25 @@
-const plugin_base = require('./plugin_base');
+/*
+ * This file is part of the Companion project
+ * Copyright (c) 2018 Bitfocus AS
+ * Authors: William Viker <william@bitfocus.io>, Håkon Nessjøen <haakon@bitfocus.io>
+ *
+ * This program is free software.
+ * You should have received a copy of the MIT licence as well as the Bitfocus
+ * Individual Contributor License Agreement for companion along with
+ * this program.
+ *
+ * You can be released from the requirements of the license by purchasing
+ * a commercial license. Buying such a license is mandatory as soon as you
+ * develop commercial activities involving the Companion software without
+ * disclosing the source code of your own applications.
+ *
+ */
+
+const common = require('./common');
 const moment = require('moment');
 
-class time_of_day extends plugin_base {
+class time_of_day extends common {
+
 	setup() {
 		this.day_format = 'ddd';
 
@@ -31,6 +49,32 @@ class time_of_day extends plugin_base {
 		];
 	}
 
+	config_desc(config) {
+		const days = config.days;
+		let day_str = days.toString();
+
+		if (days.length === 7) {
+			day_str = 'Daily';
+		}
+		else if (day_str === '1,2,3,4,5') {
+			day_str = 'Weekdays';
+		}
+		else if (day_str === '0,6') {
+			day_str = 'Weekends';
+		}
+		else {
+			try {
+				day_str = days.map(d => moment().weekday(d).format('ddd'))
+					.join(', ');
+			}
+			catch (e) {
+				day_str = 'Error';
+			}
+		}
+
+		return `<strong>${day_str}</strong>, ${config.time}`;
+	}
+
 	/**
 	 * @return {Array.<{id: number, label: string}>} List of days
 	 */
@@ -43,6 +87,14 @@ class time_of_day extends plugin_base {
 		});
 	}
 
+	get name() {
+		return 'Time of day';
+	}
+
+	get type() {
+		return 'tod';
+	}
+
 	/**
 	 * Checks if event should work
 	 * @param {number} day
@@ -52,37 +104,6 @@ class time_of_day extends plugin_base {
 		this.watch.filter(x => x.config.days.includes(day) && x.config.time === hms)
 			.forEach(x => this.scheduler.action(x.id));
 	}
-
-	get type() {
-		return 'tod';
-	}
-
-	get name() {
-		return 'Time of day';
-	}
-
-	config_desc(config) {
-		const days = config.days;
-		let day_str = days.toString();
-		if (days.length === 7) {
-			day_str = 'Daily';
-		} else if (day_str === '1,2,3,4,5') {
-			day_str = 'Weekdays';
-		} else if (day_str === '0,6') {
-			day_str = 'Weekends';
-		} else {
-			try {
-				day_str = days.map(d => moment().weekday(d).format('ddd'))
-					.join(', ');
-			} catch (e) {
-				day_str = 'Error';
-			}
-		}
-
-		return `<strong>${day_str}</strong>, ${config.time}`;
-	}
-
-
 }
 
 module.exports = time_of_day;
