+++ lib/schedule/variable.js
@@ -1,6 +1,23 @@
-const plugin_base = require('./plugin_base');
+/*
+ * This file is part of the Companion project
+ * Copyright (c) 2018 Bitfocus AS
+ * Authors: William Viker <william@bitfocus.io>, Håkon Nessjøen <haakon@bitfocus.io>
+ *
+ * This program is free software.
+ * You should have received a copy of the MIT licence as well as the Bitfocus
+ * Individual Contributor License Agreement for companion along with
+ * this program.
+ *
+ * You can be released from the requirements of the license by purchasing
+ * a commercial license. Buying such a license is mandatory as soon as you
+ * develop commercial activities involving the Companion software without
+ * disclosing the source code of your own applications.
+ *
+ */
 
-class variable extends plugin_base {
+const common = require('./common');
+
+class variable extends common {
 	setup() {
 		this.scheduler.system.on('variable_changed', this.matches.bind(this));
